Setup the 'Projects' subsection from the main section
Both the layout and the style have been set up.
The style resemble the one of the 'Trending' and 'Announcing'
sections, for uniformity. However, as this is the main content of
the webpage, some decoration images have been added as ::before
elements on the projects items, to highlight them. These images get
larger when hovering over the project item.

This subsection is responsive, too. This is achieved exploiting the
grid layout properties, using:

	grid-template-columns: repeat(auto-fill, minmax(min(
			var(--main-box-project-max-width), 100%), 1fr))

All the project items have the same size. Also, the displayed text
is clipped at a given maximum number of rows.